chore: add enums to shared folder
