+++ src/shared/util/jobs.ts
@@ -1,0 +1,11 @@
+export enum JobItemStatusEnum {
+  InProgress = 'IN_PROGRESS',
+  Success = 'SUCCESS',
+  Failure = 'FAILURE',
+}
+
+export enum JobStatusEnum {
+  InProgress = 'IN_PROGRESS',
+  Success = 'SUCCESS',
+  Failure = 'FAILURE',
+}
